+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2022-08
 
+- if not already, load full template when creating new mission
 - expose new providerMissionId
 - disable subscription stuff for now
 - adjustment to enable ui dispatch from mxm-server (via quarkus-quinoa)